Revert "Reimplement permissions fetching to do it in a single transaction (#21156)"
This reverts commit ed6212a1
